worker class refactored
